Several Jagger can use one database in the same time. Master delete information about his session only.

+++ chassis/core/src/main/java/com/griddynamics/jagger/storage/rdb/HibernateKeyValueStorage.java
@@ -32,8 +32,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Required;
 import org.springframework.orm.hibernate3.support.HibernateDaoSupport;
 
-import java.util.Collection;
-import java.util.List;
+import java.util.*;
 
 public class HibernateKeyValueStorage extends HibernateDaoSupport implements KeyValueStorage {
 
@@ -41,8 +40,25 @@ public class HibernateKeyValueStorage extends HibernateDaoSupport implements Key
 
     private int hibernateBatchSize;
 
+    private int sessionTempDataCount=50;
+
+    private String sessionId;
+
     public int getHibernateBatchSize() {
         return hibernateBatchSize;
+    }
+
+    public int getSessionTempDataCount() {
+        return sessionTempDataCount;
+    }
+
+    @Required
+    public void setSessionTempDataCount(int sessionTempDataCount) {
+        if (sessionTempDataCount <0) {
+            log.warn("Session count can't be < 0; was get {}. ", sessionTempDataCount);
+            return;
+        }
+        this.sessionTempDataCount = sessionTempDataCount;
     }
 
     @Required
@@ -57,6 +73,11 @@ public class HibernateKeyValueStorage extends HibernateDaoSupport implements Key
 
     @Override
     public void initialize() {
+    }
+
+    @Override
+    public void setSessionId(String sessionId) {
+        this.sessionId=sessionId;
     }
 
     @Override
@@ -90,8 +111,21 @@ public class HibernateKeyValueStorage extends HibernateDaoSupport implements Key
     }
 
     @Override
-    public void deleteAll(){
-        getHibernateTemplate().bulkUpdate("delete from KeyValue");
+    public void deleteAll(String sessionId) {
+        ArrayList<String> sessions = (ArrayList) getHibernateTemplate().find("Select distinct k.sessionId from KeyValue k ORDER by k.sessionId");
+        if (sessions.size() == 0)
+            return;
+        if (sessionTempDataCount == 0) {
+            log.warn("Session count limit is equal '0', all temporary data about sessions will be delete");
+            getHibernateTemplate().bulkUpdate("delete from KeyValue");
+            return;
+        }
+        List<String> sessionForDelete = Lists.newArrayList();
+        sessionForDelete.add(sessionId);
+        if (sessions.size() > sessionTempDataCount) {
+            sessionForDelete.addAll(sessions.subList(0, (sessions.size() - 1) - sessionTempDataCount));
+        }
+        getHibernateTemplate().bulkUpdate("delete from KeyValue where sessionId in (?)", sessionForDelete.toArray());
     }
 
     @SuppressWarnings("unchecked")
@@ -146,6 +180,7 @@ public class HibernateKeyValueStorage extends HibernateDaoSupport implements Key
         keyvalue.setNamespace(namespace.toString());
         keyvalue.setKey(key);
         keyvalue.setData(SerializationUtils.serialize(value));
+        keyvalue.setSessionId(sessionId);
         return keyvalue;
     }
 }
